Added model string

+++ app/src/main/java/chat/cherish/xxim/sdk/listener/UnreadListener.java
@@ -1,11 +1,11 @@
 package chat.cherish.xxim.sdk.listener;
 
 interface IUnread {
-    void onUnreadCount(int count);
+    void onUnreadCount(long count);
 }
 
 public abstract class UnreadListener implements IUnread {
     @Override
-    public void onUnreadCount(int count) {
+    public void onUnreadCount(long count) {
     }
 }
